Added DiProvider class. Updated Application class.
