+++ presentation/src/main/kotlin/com/dvinc/workouttimer/presentation/common/application/WorkoutApp.kt
@@ -12,13 +12,7 @@ import com.crashlytics.android.Crashlytics
 import com.dvinc.workouttimer.BuildConfig
 import io.fabric.sdk.android.Fabric
 import com.dvinc.workouttimer.presentation.common.timber.ReleaseTree
-import com.dvinc.workouttimer.presentation.di.component.DaggerAppComponent
-import com.dvinc.workouttimer.presentation.di.component.AppComponent
-import com.dvinc.workouttimer.presentation.di.component.ExerciseComponent
-import com.dvinc.workouttimer.presentation.di.component.NewWorkoutComponent
-import com.dvinc.workouttimer.presentation.di.component.WorkoutComponent
-import com.dvinc.workouttimer.presentation.di.component.NewExerciseComponent
-import com.dvinc.workouttimer.presentation.di.module.AppModule
+import com.dvinc.workouttimer.presentation.di.provider.DiProvider
 import com.facebook.stetho.Stetho
 import com.squareup.leakcanary.LeakCanary
 import timber.log.Timber
@@ -32,87 +26,19 @@ class WorkoutApp : Application() {
 
         private lateinit var refWatcher: RefWatcher
 
-        fun get(context: Context): WorkoutApp {
-            return context.applicationContext as WorkoutApp
-        }
-
         fun getRefWatcher(): RefWatcher {
             return refWatcher
         }
     }
 
-    lateinit var appComponent: AppComponent
-
-    private var workoutComponent: WorkoutComponent? = null
-
-    private var exerciseComponent: ExerciseComponent? = null
-
-    private var newWorkoutComponent: NewWorkoutComponent? = null
-
-    private var newExerciseComponent: NewExerciseComponent? = null
-
     override fun onCreate() {
         super.onCreate()
         WorkoutApp.context = this
-        appComponent = buildDI()
+        DiProvider.buildDi(this)
         Timber.plant(if (BuildConfig.DEBUG) Timber.DebugTree() else ReleaseTree())
         Fabric.with(this, Crashlytics())
         Stetho.initializeWithDefaults(this)
         setupLeakCanary()
-    }
-
-    fun getWorkoutComponent(): WorkoutComponent? {
-        if (workoutComponent == null) {
-            workoutComponent = appComponent.getWorkoutComponent()
-        }
-
-        return workoutComponent
-    }
-
-    fun clearWorkoutComponent() {
-        workoutComponent = null
-    }
-
-    fun getExerciseComponent(): ExerciseComponent? {
-        if (exerciseComponent == null) {
-            exerciseComponent = appComponent.getExerciseComponent()
-        }
-
-        return exerciseComponent
-    }
-
-    fun clearExerciseComponent() {
-        exerciseComponent = null
-    }
-
-    fun getNewWorkoutComponent(): NewWorkoutComponent? {
-        if (newWorkoutComponent == null) {
-            newWorkoutComponent = appComponent.getNewWorkoutComponent()
-        }
-
-        return newWorkoutComponent
-    }
-
-    fun clearNewWorkoutComponent() {
-        newWorkoutComponent = null
-    }
-
-    fun getNewExerciseComponent(): NewExerciseComponent? {
-        if (newExerciseComponent == null) {
-            newExerciseComponent = appComponent.getNewExerciseComponent()
-        }
-
-        return newExerciseComponent
-    }
-
-    fun clearNewExerciseComponent() {
-        newWorkoutComponent = null
-    }
-
-    private fun buildDI(): AppComponent {
-        return DaggerAppComponent.builder()
-                .appModule(AppModule(this))
-                .build()
     }
 
     private fun setupLeakCanary() {
